Build: Add dependency analysis plugin to example module
As per the documentation, this change follows the below step within the
'Adding a new dependency' section:

"If this is a new plugin, add it under plugins section in
example/build.gradle.kts - make sure to add .apply(false). This step is
necessary because we use the example module to generate a dependency
cache we use to speed up CI in other projects. We are able to
automatically add the library dependencies, so this step is not
necessary, but plugins need to be added manually."